Move the talk forms to a new module.
