+++ conference/forms/__init__.py
@@ -1,0 +1,2 @@
+from .forms import *  # noqa
+from .talks import *  # noqa
